Included the algorithm within the model without calling an external file
