+++ algorithm.rb
